# src/contracts/v1/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SmartSketch / 智绘学途 API",
    "version": "1.0.0",
    "description": "课程知识图谱智能构建与学习导航系统的前后端共享契约。\n\n命名基线为 ADR-008，不得引入同义别名。本文件是契约的单一真源（ADR-004）：\n实现代码、Cypher、Pydantic 模型与前端类型全部由它生成或以它为准，不得反向手改。\n\n两类状态不是 HTTP 错误，必须以 200 返回可机读状态（AGENTS.md 第 4 节、ADR-003）：\n  - 问答证据不足 → `ChatResponse.status = not_covered`\n  - 任务处理失败 → `Task.stage = failed`，附 `Task.error`\n"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "本地开发"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/courses/{cid}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/CourseId"
        }
      ],
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "课程详情",
        "operationId": "getCourse",
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "200": {
            "description": "课程详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": "学生成员读取从未发布的课程（`GRAPH_NOT_PUBLISHED`）。课程不存在与非成员同为 403 `COURSE_FORBIDDEN`，见 `specs/identity-access.md` §4.1。"
          }
        }
      }
    },
...
    "/api/v1/courses/{cid}/kp": {
      "parameters": [
        {
          "$ref": "#/components/parameters/CourseId"
        }
      ],
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "知识点列表（卡片/列表视图）",
        "description": "支撑规格验收条件 5 的卡片视图，与图谱选中态联动。",
        "operationId": "listKnowledgePoints",
        "parameters": [
          {
            "name": "chapter_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/KnowledgePointType"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "按名称与别名搜索",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "200": {
            "description": "知识点列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KnowledgePoint"
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "新增知识点（教师）",
        "description": "草稿写入须持课程写锁；超时返回 409 `COURSE_BUSY`。\n必须带至少一条来源 `sources`（ADR-035）：文本块须属于本课程，且其资料修订属于已提交任务\n（`awaiting_review` 或 `completed`）；否则 422 `VALIDATION_ERROR`，`details.fields` 指向出错的来源。\n",
        "operationId": "createKnowledgePoint",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgePointCreate"
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "201": {
            "description": "已创建，`source` 固定为 manual、`locked` 固定为 true",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgePoint"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/api/v1/courses/{cid}/kp/{kid}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/CourseId"
        },
        {
          "$ref": "#/components/parameters/KnowledgePointId"
        }
      ],
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "知识点详情与出处",
        "description": "必须返回至少一条 `source_refs`（规格验收条件 5）。",
        "operationId": "getKnowledgePoint",
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "200": {
            "description": "知识点详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgePointDetail"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": [
          "graph"
        ],
        "summary": "修改知识点（教师）",
        "description": "请求必须携带节点级 `expected_revision`；与当前 `revision` 不一致时 409 `REVISION_CONFLICT`，\n不覆盖先前写入，`details` 给出当前修订号与当前内容（ADR-035）。修改成功后修订号递增，\n`locked` 置为 true，后续自动流程不再覆盖；本接口不能解锁，解锁见 `unlockKnowledgePoint`。\n草稿写入须持课程写锁；超时返回 409 `COURSE_BUSY`。\n",
        "operationId": "updateKnowledgePoint",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgePointUpdate"
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "200": {
            "description": "已更新",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgePoint"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      },
      "delete": {
        "tags": [
          "graph"
        ],
        "summary": "删除知识点及其关系（教师）",
        "description": "从草稿删除知识点、与它相连的全部草稿关系（含关系身份）及其来源关联；共享文本块、已发布版本的副本与快照\n不变（ADR-048）。节点不存在、不可见或属于他课 → 404 `NOT_FOUND`；并发删除同一节点恰有一次 204，\n其余 404。可选 `expected_revision` 与当前修订号不一致时 409 `REVISION_CONFLICT`，不删除。\n草稿写入须持课程写锁；超时返回 409 `COURSE_BUSY`。\n",
        "operationId": "deleteKnowledgePoint",
        "parameters": [
          {
            "name": "expected_revision",
            "in": "query",
            "required": false,
            "description": "节点级乐观并发：读到的 `revision`；省略时不核对（ADR-048）",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "204": {
            "description": "已删除"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/api/v1/courses/{cid}/kp/{kid}/unlock": {
      "parameters": [
        {
          "$ref": "#/components/parameters/CourseId"
        },
        {
          "$ref": "#/components/parameters/KnowledgePointId"
        }
      ],
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "解锁知识点（教师）",
        "description": "显式解锁（规格验收条件 11，ADR-035）：`locked` 置为 false，修订号递增，此后自动流程可以更新该节点。\n须带 `expected_revision`，不一致时 409 `REVISION_CONFLICT`；节点本就未加锁时 200 原样返回，修订号不变。\n草稿写入须持课程写锁；超时返回 409 `COURSE_BUSY`。\n",
        "operationId": "unlockKnowledgePoint",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgePointUnlock"
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "200": {
            "description": "已解锁（或本就未加锁）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgePoint"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/api/v1/courses/{cid}/kp/merge": {
      "parameters": [
        {
          "$ref": "#/components/parameters/CourseId"
        }
      ],
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "合并重复知识点（教师）",
        "description": "被合并节点的名称并入 `aliases`，关系与来源取并集后迁移到主节点（ADR-047）：入边与出边改接到主节点，\n同类型同端点的关系去重（主节点原有的关系保留字段，贡献与来源取并集），两端都在本次合并内的关系删除；\n迁移后的 `PREREQUISITE` 执行 DAG 环检测，成环返回 409 `CYCLE_DETECTED` 并给出 `details.cycle`。\n主节点修订号加 1 并加锁，被合并节点从草稿删除、记入主节点的合并谱系。可选 `expected_revisions`\n与当前修订号不一致时 409 `REVISION_CONFLICT`。节点不存在或不可见时 422（`details.fields` 的 `reason`\n为 `not_found`）。草稿写入须持课程写锁；超时返回 409 `COURSE_BUSY`。任何失败都不部分提交。\n",
        "operationId": "mergeKnowledgePoints",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeRequest"
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "200": {
            "description": "合并后的主节点",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgePoint"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "令牌载荷含 `sub`（用户 ID）与 `role`（teacher / student）。\n`role` 仅作前端路由提示；服务端不据此授权，课程内权限每次回查成员关系。\n"
      },
      "eventTicket": {
        "type": "apiKey",
        "in": "query",
        "name": "ticket",
        "description": "任务 SSE 使用的一次性票据；申领与核销规则见 `specs/identity-access.md` §5。"
      }
    },
    "parameters": {
      "CourseId": {
        "name": "cid",
        "in": "path",
        "required": true,
        "description": "课程 ID，所有查询的第一隔离条件",
        "schema": {
          "type": "string"
        }
      },
...
      "KnowledgePointId": {
        "name": "kid",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
      "Unauthenticated": {
        "description": "未认证或令牌失效",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "角色不允许，或跨课程访问（`COURSE_FORBIDDEN`）",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在，或学生读取未发布图谱（`GRAPH_NOT_PUBLISHED`）",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "状态冲突，如 `CYCLE_DETECTED`、`NODE_LOCKED`、`TASK_NOT_CANCELLABLE`、`REVISION_CONFLICT`、`COURSE_BUSY`",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "请求体校验失败（`VALIDATION_ERROR`）",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "description": "面向用户的可读说明"
          },
          "details": {
            "type": "object",
            "description": "结构化补充，例如 `CYCLE_DETECTED` 的 `cycle` 节点链路",
            "additionalProperties": true
          }
        }
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "teacher",
          "student"
        ]
      },
...
      "CourseStatus": {
        "type": "string",
        "description": "draft 未发布；published 已发布；revising 已发布后再修改，学生仍见上一版",
        "enum": [
          "draft",
          "published",
          "revising"
        ]
      },
      "Course": {
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "my_role",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/CourseStatus"
          },
          "my_role": {
            "$ref": "#/components/schemas/Role"
          },
          "teacher_id": {
            "type": "string"
          },
          "kp_count": {
            "type": "integer",
            "minimum": 0
          },
          "published_version": {
            "type": [
              "integer",
              "null"
            ],
            "description": "学生可见的版本号；从未发布为 null"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "KnowledgePointType": {
        "type": "string",
        "description": "对应 S2 表 6.3 的五类知识点",
        "enum": [
          "concept",
          "theorem",
          "formula",
          "method",
          "example"
        ]
      },
      "KnowledgePointStatus": {
        "type": "string",
        "description": "审核状态；low_confidence 项进入审核队列",
        "enum": [
          "draft",
          "low_confidence",
          "approved",
          "rejected"
        ]
      },
      "NodeSource": {
        "type": "string",
        "enum": [
          "ai",
          "manual"
        ]
      },
...
      "KnowledgePoint": {
        "type": "object",
        "required": [
          "id",
          "course_id",
          "name",
          "type",
          "definition",
          "level",
          "confidence",
          "status",
          "source",
          "locked",
          "revision"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "course_id": {
            "type": "string"
          },
          "chapter_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "aliases": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "type": {
            "$ref": "#/components/schemas/KnowledgePointType"
          },
          "definition": {
            "type": "string"
          },
          "importance": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "difficulty": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "level": {
            "type": "integer",
            "minimum": 0,
            "description": "前置关系图中的最长前置路径长度，用于层次布局与路径排序"
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "status": {
            "$ref": "#/components/schemas/KnowledgePointStatus"
          },
          "source": {
            "$ref": "#/components/schemas/NodeSource"
          },
          "locked": {
            "type": "boolean",
            "description": "true 时不被后续自动抽取/融合流程覆盖"
          },
          "revision": {
            "type": "integer",
            "minimum": 1,
            "description": "节点级乐观并发修订号；不同于课程 draft_revision。"
          },
          "source_refs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceRef"
            }
          }
        }
      },
      "KnowledgePointDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/KnowledgePoint"
          },
          {
            "type": "object",
            "required": [
              "source_refs"
            ],
            "properties": {
              "source_refs": {
                "type": "array",
                "minItems": 1,
                "items": {
                  "$ref": "#/components/schemas/SourceRef"
                }
              },
              "prerequisites": {
                "type": "array",
                "description": "直接前置知识点",
                "items": {
                  "$ref": "#/components/schemas/KnowledgePointRef"
                }
              },
              "successors": {
                "type": "array",
                "description": "直接后继知识点",
                "items": {
                  "$ref": "#/components/schemas/KnowledgePointRef"
                }
              },
              "related": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/KnowledgePointRef"
                }
              }
            }
          }
        ]
      },
      "KnowledgePointRef": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/KnowledgePointType"
          }
        }
      },
      "KnowledgePointCreate": {
        "type": "object",
        "required": [
          "name",
          "type",
          "definition",
          "sources"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "sources": {
            "type": "array",
            "minItems": 1,
            "description": "新建知识点的来源，至少一条（ADR-035）",
            "items": {
              "$ref": "#/components/schemas/KnowledgePointSourceInput"
            }
          },
          "aliases": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "type": {
            "$ref": "#/components/schemas/KnowledgePointType"
          },
          "definition": {
            "type": "string"
          },
          "chapter_id": {
            "type": "string"
          },
          "importance": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "difficulty": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          }
        }
      },
...
      "KnowledgePointUnlock": {
        "type": "object",
        "required": [
          "expected_revision"
        ],
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "KnowledgePointUpdate": {
        "description": "必带 expected_revision，且至少提供一个可修改字段。",
        "allOf": [
          {
            "type": "object",
            "required": [
              "expected_revision"
            ],
            "additionalProperties": false,
            "properties": {
              "expected_revision": {
                "type": "integer",
                "minimum": 1
              },
              "name": {
                "type": "string"
              },
              "aliases": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "type": {
                "$ref": "#/components/schemas/KnowledgePointType"
              },
              "definition": {
                "type": "string"
              },
              "importance": {
                "type": "number",
                "minimum": 0,
                "maximum": 1
              },
              "difficulty": {
                "type": "number",
                "minimum": 0,
                "maximum": 1
              },
              "status": {
                "$ref": "#/components/schemas/KnowledgePointStatus"
              }
            }
          },
          {
            "anyOf": [
              {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              },
              {
                "type": "object",
                "required": [
                  "aliases"
                ],
                "properties": {
                  "aliases": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "$ref": "#/components/schemas/KnowledgePointType"
                  }
                }
              },
              {
                "type": "object",
                "required": [
                  "definition"
                ],
                "properties": {
                  "definition": {
                    "type": "string"
                  }
                }
              },
              {
                "type": "object",
                "required": [
                  "importance"
                ],
                "properties": {
                  "importance": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 1
                  }
                }
              },
              {
                "type": "object",
                "required": [
                  "difficulty"
                ],
                "properties": {
                  "difficulty": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 1
                  }
                }
              },
              {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "$ref": "#/components/schemas/KnowledgePointStatus"
                  }
                }
              }
            ]
          }
        ]
      },
      "MergeRequest": {
        "type": "object",
        "required": [
          "primary_id",
          "merged_ids"
        ],
        "additionalProperties": false,
        "properties": {
          "primary_id": {
            "type": "string",
            "description": "保留的主节点；其名称成为主名"
          },
          "merged_ids": {
            "type": "array",
            "minItems": 1,
            "description": "并入主节点的知识点，不得含主节点，不得重复",
            "items": {
              "type": "string"
            }
          },
          "expected_revisions": {
            "type": "object",
            "description": "可选的节点级乐观并发：键为本次合并涉及的知识点 ID（主节点或被合并节点），值为读到的 `revision`。\n任一不一致时 409 `REVISION_CONFLICT`，不写入（ADR-047）。\n",
            "additionalProperties": {
              "type": "integer",
              "minimum": 1
            }
          }
        }
      },
...
    }
  }
}